backend/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
transformers
torch
